Feat: OrderController, Customer free to access and create order, getAllOrders, Only staff can change the status of the order

+++ frontend/frontend/src/main/java/com/example/SelfOrderingRestaurant/Dto/Response/OrderResponseDTO/GetAllOrdersResponseDTO.java
@@ -18,7 +18,7 @@ public class GetAllOrdersResponseDTO {
     private String customerName;
     private int tableNumber;
     private String status;
-    private Long totalAmount;
+    private BigDecimal totalAmount;
     private String paymentStatus;
     private List<OrderItemDTO> items;
 }
